post_delete에서 작성자와 요청자가 같을 때만 삭제

+++ django/blog/views.py
@@ -38,5 +38,8 @@ def post_add(request):
 def post_delete(request, pk):
     if request.method == 'POST':
         post = Post.objects.get(pk=pk)
-        post.delete()
-    return redirect('post-list')
+
+        if request.user == post.author:
+            post.delete()
+            return redirect('post-list')
+        return redirect('post-detail', pk=post.pk)
